20250807 Autumn-beginning

+++ inc/cpp/Device/SD.hpp
@@ -76,7 +76,7 @@ namespace uni {
 		virtual bool Write(stduint BlockIden, const void* Sors);
 		virtual stduint getUnits() { return _TODO 0; }
 		//
-		virtual byte operator[](uint64 bytid) { return _TODO false; }// byte read
+		virtual int operator[](uint64 bytid) { return _TODO false; }// byte read
 		// ---- RuptTrait ----
 		virtual void setInterrupt(Handler_t _func) const override { _TODO }
 		virtual void setInterruptPriority(byte preempt, byte sub_priority) const override { _TODO }
